+++ website/src/examples-sidebar.js
@@ -55,7 +55,7 @@ const sidebars = {
     {
       type: 'category',
       label: '@deck.gl-community/three',
-      items: ['three/wild-forest']
+      items: ['three/seasonal-farm']
     },
     {
       type: 'category',